requirements.txt
pyyaml
pandas
openai
tqdm